requirements.txt
asttokens==2.2.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
Flask==1.1.1
numpy==1.21.6
scikit_learn==0.23.1
Werkzeug==0.15.5
Jinja2==2.10.1
itsdangerous==1.1.0
MarkupSafe==1.1.1
pandas==1.3.5
scipy==1.7.3
gunicorn==20.1.0
kiwisolver==1.4.4
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
#pywin32==305
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
torch==1.13.0
torchvision==0.14.0
tornado==6.2
traitlets==5.5.0
typing-extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
python-utils==3.4.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
git+https://github.com/Rapptz/discord.py
ffmpeg